refactor(tests): one definition of the XRAS API auth fixtures
## Summary

`test_xras_access.py` held `XRAS_PW`, `_basic`, `_reset_db_key_cache`, `xras_keys`
and `xras_client` privately. Two new HTTP-tier modules need all five, and copying
them would re-create exactly the duplication the pre-cutover tidy-up removed from
this tier a few commits ago.

They move to `tests/xras_helpers.py`, which already exists for this — it holds
`committing`, `load_fixture`, `wire_resource` and `txns_for` for the same reason.

The comment that travels with them is the point: a `make_api_credentials` row is
**invisible** to an HTTP request, because routes read Flask-SQLAlchemy's
`db.session` on a separate connection that only sees committed rows. So the DB-key
loader is monkeypatched rather than seeded. That reasoning was written down once
and would have been paraphrased twice more.

`reset_db_key_cache` stays `autouse`, so importing it is the whole of the wiring.

### Test Results

`pytest tests/api/test_xras_access.py -q` → **113 passed**, with no test-body edits
— only the import block changed.
